Added SortFullName link to NavBar
Having Issues displaying the sorted names. Get this error: The character encoding of the HTML document was not declared. The document will render with garbled text in some browser configurations if the document contains characters from outside the US-ASCII range. The character encoding of the page must be declared in the document or in the transfer protocol.
Tried to fix it but couldn't figure it out. Didn't commit, because we didn't change anything.